requirements.txt
pandas
numpy
scikit-learn
mysql-connector-python
python-dotenv
nltk
SpaCy
# pysql
pymysql
flask
joblib
